requirements.txt
streamlit
streamlit-lottie
tensorflow
numpy
Pillow
joblib
scikit-learn
lightgbm
requests
matplotlib